Moved averaging functions from the bin script average.py to a new module. Now glocal alignment is no longer importing functions from a bin script.
