+++ pytom/alignment/__init__.py
@@ -1,2 +1,2 @@
 __all__ = ["structures","alignmentFunctions","ExMaxAlignment","MultiRefAlignment","MultiRefStructures","FRMAlignment",
-          "localOptimization","GLocalSampling", 'alignmentStructures']
+          "localOptimization","GLocalSampling", "alignmentStructures", "averaging"]
